Add human-game-client-server cycle diagram
...and `dot` source. The idea is to show input/output flows,
with every edge between nodes representing a transformation of
some sort (e.g. physical movement of hands into data streams
from keyboard and mouse). Every input/output may be associated
with a distribution and parameters for its arrival/departure
process. To find the end-to-end lag, sum the lags along every
edge in the path or cycle of interest.