Add two pieces of related work
One was suggested by PQS Reviewer 1 and is a great match, particularly
in terms of QoE (though orthorgonal with respect to lag, and lacking
modelling).

The other is a 1994 paper found via the first. It contains a few
interesting ideas, e.g. given the lag, how does the optimal timing
of displaying the scene and sampling user input look like, so as to
minimize the user's performance in the test task.